+++ VegetableChef.cpp
@@ -48,7 +48,7 @@ void VegetableChef::receiveOrder(FoodItem* foodItem, int tableNumber, int custID
         sleep(1);
         this->addMeal(createMain());
     }
-    if(plate->getMeals().size() == numOfItems)
+    if((int)plate->getMeals().size() == numOfItems)
     {
         mediator->notifyPlateReady(plate);
     }
